PoC: create a 'pr number' type for datatables
The basic code ought to work, but does not do what it's supposed to be:
the first column is a HTML tag with the data embedded; it would be best to
re-use the code for html-num... or use a custom renderer instead.

The second issue is that the way I pasted the script is also wrong,
meaning there's a syntax error and the datatable setup is deactivated.
Should be easy to fix.